Remove delete btn in question view & answer idx not asked by currentuser

+++ frontend/components/answer/answer_index_item.jsx
@@ -8,16 +8,22 @@ export default class AnswerIndexItem extends React.Component {
 
   render() {
     let topicName = this.props.topics[this.props.question.topic_id].name;
+    let deleteButton;
+    if (this.props.currentUser.id === this.props.question.author_id) {
+      deleteButton = <span onClick={
+          () => this.props.deleteQuestion(
+            this.props.question.id,
+            this.props.question.author_id
+          )
+        }
+        className='question-delete-button'>&times;</span>;
+    } else {
+      deleteButton = null;
+    }
 
     return (
       <div className='answer-item-box'>
-        <span onClick={
-            () => this.props.deleteQuestion(
-              this.props.question.id,
-              this.props.question.author_id
-            )
-          }
-          className='question-delete-button'>&times;</span>
+        { deleteButton }
         <div className='answer-item-header'>
           Question asked · <Link to={`/topics/${this.props.question.topic_id}`}
           className='answer-topic-link'>{ topicName }</Link>
